Attempt to resolve a build issue in GitHub CI
Attempting to publish the 3.2.8 release, I discovered that the build failed in
CI because of an ugly class-cast exception in Xerces. It’s unclear what causes
this. Web searching reveals some suggestions, but many involve fussing with the
loader and pointing it at private classes in the com.sun packages. If that even
still works, it’s ugly. My best hypothesis is that it’s caused by mixing
different versions of Xerces on the class path.

The changes in this PR seem to make the problem go away. But they’re not very
satisfying. You can’t point at this PR and say, “oh, yes, I see how that fixes
it.” But it fixes it and I want to rewrite all the build scripts in Kotlin
anyway, so I’m just pressing on.